requirements.txt
streamlit==1.37.0
openai==1.40.0
chromadb==0.5.5
python-dotenv==1.0.1
langchain-community==0.2.10
langchain-text-splitters==0.2.2
docx2txt==0.8
pypdf==4.3.1